[#26] feat: footer를 별도 컴포넌트로 분리 및 적용
